change Float64List to Float32List

+++ packages/flame/lib/src/game/notifying_vector2.dart
@@ -185,5 +185,5 @@ class NotifyingVector2 extends Vector2 with ChangeNotifier {
   }
 
   @override
-  Float64List get storage => super.storage.asUnmodifiableView();
+  Float32List get storage => super.storage.asUnmodifiableView();
 }
